refactor(ledger): extract ledger-common module; move Transaction, TransactionValidator, ExceptionMessages and update poms, modules and imports
