+++ src/ledger/balancereader/pom.xml
@@ -73,6 +73,11 @@
 
     <dependencies>
         <dependency>
+            <groupId>com.google.cloud.anthos.samples</groupId>
+            <artifactId>ledger-common</artifactId>
+            <version>1.0</version>
+        </dependency>
+        <dependency>
             <groupId>org.springframework.boot</groupId>
             <artifactId>spring-boot-starter-web</artifactId>
         </dependency>
