+++ src/ledger/ledger-common/src/main/java/anthos/samples/bankofanthos/ledger/common/Transaction.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package anthos.samples.bankofanthos.balancereader;
+package anthos.samples.bankofanthos.ledger.common;
 
 import java.util.Date;
 
@@ -24,6 +24,7 @@ import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
 import jakarta.persistence.Id;
 import jakarta.persistence.Table;
+import jakarta.persistence.Transient;
 
 import org.hibernate.annotations.CreationTimestamp;
 
@@ -61,6 +62,12 @@ public final class Transaction {
     @JsonProperty("timestamp")
     private Date timestamp;
 
+    // UUID is used for preventing duplicate requests from client
+    // Do not persist to database
+    @Transient
+    @JsonProperty("uuid")
+    private String requestUuid;
+
     private static final double CENTS_PER_DOLLAR = 100.0;
 
     public long getTransactionId() {
@@ -86,6 +93,19 @@ public final class Transaction {
     public Integer getAmount() {
         return amount;
     }
+
+    public Date getTimestamp() {
+        return timestamp;
+    }
+
+    public String getRequestUuid() {
+        if (requestUuid == null) {
+            return "";
+        } else {
+            return requestUuid;
+        }
+    }
+
     /**
      * String representation.
      *
