+++ src/ledger/ledgerwriter/src/main/java/anthos/samples/bankofanthos/ledgerwriter/LedgerWriterApplication.java
@@ -28,6 +28,7 @@ import org.apache.logging.log4j.Logger;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.actuate.autoconfigure.tracing.zipkin.ZipkinAutoConfiguration;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.autoconfigure.domain.EntityScan;
 import org.springframework.context.annotation.Bean;
 import org.springframework.web.client.RestTemplate;
 
@@ -37,6 +38,7 @@ import org.springframework.web.client.RestTemplate;
  * Microservice to accept new transactions for the bank ledger.
  */
 @SpringBootApplication(exclude = ZipkinAutoConfiguration.class)
+@EntityScan(basePackages = "anthos.samples.bankofanthos.ledger.common")
 public class LedgerWriterApplication {
 
     private static final Logger LOGGER =
